Extract AlignArguments out from AlignParameters
